// src/app/shared/model/reporting.model.ts
export interface ITag {
    nbCA: number;
    UsrModif: string;
    nbOK: number;
    nbExe: number;
    description: string;
    DateCreated: string;
    DateEndQueue: string;
    nbKO: number;
    nbQU: number;
    DateModif: string;
    nbWE: number;
    nbNA: number;
    ciResult: string;
    nbNE: number;
    ciScoreThreshold: number;
    campaign: string;
    UsrCreated: string;
    nbPE: number;
    ciScore: number;
    id: number;
    tag: string;
    nbQE: number;
    nbFA: number;
    nbExeUsefull: number;
}